EDI/requirements.txt
# Core dependencies (existing)
flask>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# New dependencies for upgraded features
yfinance>=0.2.28
schedule>=1.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# Optional but recommended
python-dotenv>=1.0.0
requests>=2.31.0
